% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCContour.R
\name{cindex_auc_internal_imports}
\alias{cindex_auc_internal_imports}
\title{Internal imports for C-index / AUC contour functions}
\description{
Internal imports for C-index / AUC contour functions
}
\keyword{internal}
