% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectContour.R
\name{effect_internal_imports}
\alias{effect_internal_imports}
\title{Internal imports for effect-size contour functions}
\description{
Internal imports for effect-size contour functions
}
\keyword{internal}
