% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectContour.R
\name{rowDeri}
\alias{rowDeri}
\title{Compute Row-wise Derivative over Covariate}
\usage{
rowDeri(mat, cov)
}
\arguments{
\item{mat}{A matrix of cumulative hazard.}

\item{cov}{A numeric vector of covariate values.}
}
\value{
A matrix of row-wise derivatives.
}
\description{
Compute Row-wise Derivative over Covariate
}
